refactor: use `self` instead of using explicit arguments in methods
Moved `system` out of `self`. The reason is that some variables depend on `system` (`NetworkWidget` in particular), and `system` is mutated. This means `self` must be `mut`. Then we must use `mut self` in methods. But we can't, because we can only use `mut self` once at a time.

Refactor `update_static_props()` into a non-static call.

Rename `sysinfo_system` to `system` to be short.